test(rig): move scoped-registry rig add test to the integration lane
The new end-to-end `gc rig add --include <owner>/<pack>` test adds three
t.Setenv calls, which trips the cmd/gc+untagged environment census
(internal/testpolicy/resourcecensus) and turns Integration /
packages-core-2-of-4 red. Move the test behind //go:build integration
rather than raising the census baseline, and register it in the cmd/gc
integration manifest so it still runs in CI.

Also correct the `rig add --include` help text: a "./" prefix does not
always mean a directory — "./gastown" still canonicalizes to the bundled
source. Only the registry rung is gated on the token declaring a path.
docs/reference/cli.md is regenerated, not hand-edited.
