+++ scripts/test-integration-shard
@@ -112,6 +112,7 @@ cmd_gc_integration_tests=(
   TestControllerDiscoversAddedCronOrderWithoutRestart
   TestManagedBdRigProviderStoreRecoversAfterHardKillPortRebind
   TestPhase2WorkerCoreRealTransportProof
+  TestRigAddIncludeResolvesScopedRegistryPackName
 )
 
 join_regex() {
